MAINT: Moved setup code to _algorithm_setup
Moved the setup code for most algorithms to
_algorithm_setup. Kept morphology's setup in
morphological to allow users to call the optimize_window function.
